+++ tests/phpunit/cases/template/tags/random.php
@@ -1,14 +1,18 @@
 <?php
-class Template_Tags_Random_TestCase extends WP_UnitTestCase {
+namespace Tests\Template\Tags;
+
+class Random_TestCase extends \WP_UnitTestCase {
   public function test_template_tags_random_defaults() {
-      srand( 0 );
-      $this->assertEquals( '99', tangible_template( '<Random />' ), 'failed with seed 0' );
-      $this->assertEquals( '19', tangible_template( '<Random />' ), 'failed with seed 0' );
+    $result = tangible_template( '<Random />' );
+    $this->assertEquals( true, is_numeric($result) );
+    // Another number
+    $result = tangible_template( '<Random />' );
+    $this->assertEquals( true, is_numeric($result) );
   }
 
   public function test_template_tags_random_negative() {
-      srand( 0 );
-      $this->assertEquals( '-100', tangible_template( '<Random from=-100 to=-99/>' ), 'failed with seed 0' );
-      $this->assertEquals( '-99', tangible_template( '<Random from=-100 to=-99/>' ), 'failed with seed 0' );
+    $result = tangible_template( '<Random from=-100 to=-99/>' );
+    $this->assertEquals( true, $result >= -100, 'from -100' );
+    $this->assertEquals( true, $result <= -99, 'to -99' );
   }
 }